requirements.txt
Flask=3.1.1
python-docx=1.2.0
easyocr=1.7.2
requests=2.32.4
python-dotenv=1.1.1
pytesseract=0.3.13
pillow=11.3.0